requirements.txt
streamlit==1.48.1
pdfplumber==0.11.7
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.9
google-generativeai
faiss-cpu==1.12.0
tiktoken==0.11.0